fix bug product add ad guard for protect sent empty image"

+++ src/controller/product.js
@@ -114,12 +114,11 @@ module.exports.createProduct = async (req, res, next) => {
       throw new CustomError("name has been exist", "WRONG_INPUT", 400);
     }
     const productData = req.body;
-    // for post man
     productData.serieId = +req.body.serieId;
     productData.groupId = +req.body.groupId;
     productData.price = +req.body.price;
     productData.stockQuantity = +req.body.stockQuantity;
-    productData.launchDate = new Date(); // select date from front
+    productData.launchDate = new Date(req.body.launchDate); // select date from front
 
     let photoProduct = [];
     let photoLinkProduct = [];
@@ -131,41 +130,55 @@ module.exports.createProduct = async (req, res, next) => {
     // // upload cloudinary
 
     // /***cover*****/
-    const coverLocalPath = req.files.coverProduct[0].path;
-    const coverPath = await utils.cloudinary.upload(coverLocalPath);
-    await repo.product.createCover(newProduct.id, coverPath);
+    if (req.files.coverProduct) {
+      const coverLocalPath = req.files.coverProduct[0].path;
+      const coverPath = await utils.cloudinary.upload(coverLocalPath);
+      await repo.product.createCover(newProduct.id, coverPath);
+    }
 
     // /**product */
-    const productPath = req.files.imageProduct.map(async (image) =>
-      utils.cloudinary.upload(image.path)
-    );
-    photoLinkProduct = await Promise.all(productPath);
-    // update link on table product image
-    const updateLinkProduct = photoLinkProduct.map((link) => {
-      const data = {
-        productId: newProduct.id,
-        images: link,
-      };
-      return repo.product.createImageProduct(data);
-    });
-    photoProduct = await Promise.all(updateLinkProduct);
+    if (req.files.imageProduct) {
+      const productPath = req.files.imageProduct.map(async (image) =>
+        utils.cloudinary.upload(image.path)
+      );
+      photoLinkProduct = await Promise.all(productPath);
+      // update link on table product image
+      const updateLinkProduct = photoLinkProduct.map((link) => {
+        const data = {
+          productId: newProduct.id,
+          images: link,
+        };
+        return repo.product.createImageProduct(data);
+      });
+      photoProduct = await Promise.all(updateLinkProduct);
+    }
 
     // update poster
 
     //upload cloudinary
     let posterPath = [];
     for (let i = 1; i <= 5; i++) {
-      posterPath.push(req.files[`poster${i}`][0].path);
+      if (req.files[`poster${i}`]) {
+        posterPath.push(req.files[`poster${i}`][0].path);
+      } else {
+        posterPath.push(null);
+      }
     }
 
     const linkAllPoster = await Promise.all(
-      posterPath.map((path) => utils.cloudinary.upload(path))
+      posterPath.map((path) =>
+        path === null ? null : utils.cloudinary.upload(path)
+      )
     );
 
     const data = {};
     data.productId = newProduct.id;
-    for (let i = 0; i < linkAllPoster.length; i++) {
-      data[`posters${i + 1}`] = linkAllPoster[i];
+    for (let i = 1; i <= 5; i++) {
+      if (linkAllPoster[i]) {
+        data[`posters${i}`] = linkAllPoster[i];
+      } else {
+        data[`posters${i}`] = null;
+      }
     }
 
     await repo.product.createPoster(data);
@@ -175,12 +188,18 @@ module.exports.createProduct = async (req, res, next) => {
     console.log(err);
     next(err);
   } finally {
-    for (image of req.files.imageProduct) {
-      fs.unlink(image.path);
+    if (req.files.imageProduct) {
+      for (image of req.files.imageProduct) {
+        fs.unlink(image.path);
+      }
     }
-    fs.unlink(req.files.coverProduct[0].path);
+    if (req.files.coverProduct) {
+      fs.unlink(req.files.coverProduct[0].path);
+    }
     for (let i = 1; i <= 5; i++) {
-      fs.unlink(req.files[`poster${i}`][0].path);
+      if (req.files[`poster${i}`]) {
+        fs.unlink(req.files[`poster${i}`][0]?.path);
+      }
     }
   }
 };
@@ -196,6 +215,7 @@ module.exports.editProduct = async (req, res, next) => {
     if (searchNameDuplicate.length > 1) {
       throw new CustomError("Product name is duplicate", "WRONG_INPUT", 400);
     }
+
     const data = req.body;
     const updateResult = await repo.product.editProduct(prodId, data);
     res.status(200).json({ updateResult });
@@ -250,6 +270,17 @@ module.exports.deleteProduct = async (req, res, next) => {
     }
 
     res.status(200).json({ message: "In Active success" });
+  } catch (err) {
+    console.log(err);
+    next(err);
+  }
+};
+
+module.exports.deleteSoft = async (req, res, next) => {
+  try {
+    const productId = +req.params.productId;
+    await repo.product.deleteProductSoft(productId);
+    res.status(200).json({ message: "inactive product success" });
   } catch (err) {
     console.log(err);
     next(err);
